// src/app/State/product-entity/product-entity-types/product-entity-types.ts
export enum ProductEntityActionsType {
  LOAD_PRODUCT_ENTITY = '[Product Page]Load Product',
  LOAD_PRODUCT_ENTITY_SUCCESS = '[Product Page]Load Product Success',
  LOAD_PRODUCT_ENTITY_FAIL = '[Product Page]Load Product Fail',
  ADD_PRODUCT_ENTITY = '[Product Page]Add Product',
  ADD_PRODUCT_ENTITY_SUCCESS = '[Product Page]Add Product Success',
  UPDATE_PRODUCT_ENTITY = '[Product Page]Update Product',
  UPDATE_PRODUCT_ENTITY_SUCCESS = '[Product Page]Update Product Success',
  DELETE_PRODUCT_ENTITY = '[Product Page]Delete Product',
  DELETE_PRODUCT_ENTITY_SUCCESS = '[Product Page]Delete Product Success',
  GET_PRODUCT_ENTITY = '[Product Page]Get Product',
  GET_PRODUCT_ENTITY_SUCCESS = '[Product Page]Get Product Success',
  OPEN_POPUP_ENTITY = '[Product Page]Open Popup',
}